+++ Photo Club Hub/ViewModel/Organization/Organization.swift
@@ -142,7 +142,7 @@ extension Organization {
         for lang in Locale.preferredLanguages {
             let langID = lang.split(separator: "-").first?.uppercased() ?? "EN"
             // now check if one of the user's preferences is available for this Remark
-            for localRemark in localizedRemarks where localRemark.language.isoCodeAllCaps == langID {
+            for localRemark in localizedRemarks where localRemark.language.isoCode == langID {
                 if localRemark.localizedString != nil {
                     return localRemark.localizedString!
                 }
@@ -150,7 +150,7 @@ extension Organization {
         }
 
         // second choice: most people speak English, at least let's pretend that is the case ;-)
-        for localizedRemark in localizedRemarks where localizedRemark.language.isoCodeAllCaps == "EN" {
+        for localizedRemark in localizedRemarks where localizedRemark.language.isoCode == "EN" {
             if localizedRemark.localizedString != nil {
                 return localizedRemark.localizedString!
             }
@@ -158,7 +158,7 @@ extension Organization {
 
         // third choice: use any translation available for this expertise
         if localizedRemarks.first != nil, localizedRemarks.first!.localizedString != nil {
-            return "\(localizedRemarks.first!.localizedString!) [\(localizedRemarks.first!.language.isoCodeAllCaps)]"
+            return "\(localizedRemarks.first!.localizedString!) [\(localizedRemarks.first!.language.isoCode)]"
         }
 
         // otherwise display an error message instead of a real remark
@@ -264,7 +264,7 @@ extension Organization {
             modified = true }
 
         if let contactEmail = optionalFields.contactEmail, self.contactEmail != contactEmail {
-            self.contactEmail = contactEmail
+            self.contactEmail = contactEmail // TODO somehow contactEmail for IndividueelBO is "" rather than nil
             modified = true }
 
         if let fotobondNumber = optionalFields.fotobondNumber, self.fotobondNumber != fotobondNumber {
@@ -279,7 +279,7 @@ extension Organization {
             let isoCode: String? = localizedRemark["language"].stringValue.uppercased() // e.g. "NL", "DE" or "PDC"
             let localizedRemarkNewValue: String? = localizedRemark["value"].stringValue
 
-            if isoCode != nil && localizedRemarkNewValue != nil { // nil could happens if JSON file not schema compliant
+            if isoCode != nil && localizedRemarkNewValue != nil { // nil could occur if JSON file isn't schema compliant
                 let language = Language.findCreateUpdate(context: bgContext,
                                                          isoCode: isoCode!) // find or construct the remark's Language
                 // language updates doesn't set modified flag
@@ -299,7 +299,7 @@ extension Organization {
 				try bgContext.save() // persist modifications in PhotoClub record
  			} catch {
                 print("Error: \(error)")
-                ifDebugFatalError("Update failed for club or museum \"\(fullName)\"",
+                ifDebugFatalError("Update failed for organization \"\(fullName)\"",
                                   file: #fileID, line: #line) // likely deprecation of #fileID in Swift 6.0
                 // in release mode, if .save() fails, just continue
                 return false