+++ Photo Club Hub/ViewModel/Photographer/Expertise.swift
@@ -46,6 +46,7 @@ extension Expertise {
                                             ) -> Expertise {
 
         // execute fetchRequest to get expertise object for id=id. Query could return multiple - but shouldn't.
+        let id = id.canonicalCase
         let fetchRequest: NSFetchRequest<Expertise> = Expertise.fetchRequest()
         let predicateFormat: String = "id_ = %@" // avoid localization
         fetchRequest.predicate = NSPredicate(format: predicateFormat, argumentArray: [id])
@@ -269,13 +270,13 @@ extension Expertise {
         for lang in Locale.preferredLanguages {
             let langID = lang.split(separator: "-").first?.uppercased() ?? "EN"
             // now check if one of the user's preferences is available for this Remark
-            for localizedExpertise in localizedExpertises where localizedExpertise.language.isoCodeAllCaps == langID {
+            for localizedExpertise in localizedExpertises where localizedExpertise.language.isoCode == langID {
                 return LocalizedExpertiseResult(localizedExpertise: localizedExpertise, id: self.id)
             }
         }
 
         // second choice: most users can speak English, at least let's assume that is the case ;-)
-        for localizedExpertise in localizedExpertises where localizedExpertise.language.isoCodeAllCaps == "EN" {
+        for localizedExpertise in localizedExpertises where localizedExpertise.language.isoCode == "EN" {
             return LocalizedExpertiseResult(localizedExpertise: localizedExpertise, id: self.id)
         }